Updated 3.4.1
Add 3D Buildings is Easier - When editing a 3D Community Scene, we now have a new menu item called simply Add 3D Building. It uses the same easy process as adding a 3D Thing and is right above it in the Edit Community menu in Admin mode. This replaces the menu option that was labeled Add 3D Building in 3D Community. We also removed the related couple of functions that are no longer needed.