Updated 3.4.1
Removed Column UserName from Users - We have been phasing out username in the users table for userid as a varchar(16) and displayname as a varchar(255). The userid is a random varchar and the display name more resembles the former use of username. Display name allows duplicates, but the userid is a unique primary key field for the table.

Added code to remove username Field from users Table - In the PHP tables class, we added code to remove the username field. This will take place on your update to 3.4.1. It checks to see if it exists and removes it. This will help with consistency and making sure the userid and displayname are used instead.

Removed Username from Connect Files - There were 3 remaining files referencing the username in the Connect folder; communityshare.php, buildingshare.php, and thingshare.php. The reference has been removed.