Updated 3.4.2
Added Check Host function to API Keys Connect - In the connect folder apikeys.php file, we added a function to confirm if an API Key and Secret are Valid. There are three possible outcomes; Valid key, Invalid Key, and Waiting on Approval. This process is designed to be called from the WordPress WalkTheWeb Plugin to confirm if a key is valid before using it to allow the WordPress WalkTheWeb Plugin to create new 3D Websites on your WalkTheWeb 3D Internet Server. This process requires a request, approval, and retest before that stage is accepted.