Updated 3.4.2
Added voicechat.js to 3D Internet Plugin - On the 3D Internet Plugin, we added a voicechat channel using voicechat.js. It is based on the chat functionality and will be the container for streaming voice to the 3D Scenes. With this, we also starting added support for socket.io-stream and recordRTC. On the 3D WalkTheWeb Hub (3dnet.walktheweb.network) server we evoked a new channel specific for voicechat.

New Variables and Settings for Voice Chat - On the 3D Internet settings page in admin, we added the ability to turn on Voice Chat. (work in progress for testing). This includes new 3D Internet global variables, form fields, and related scripts.

Groundwork for Streaming Video ad Audio - We started putting in the groundwork for streaming video and audio. This includes a test link in the Admin - 3D Internet - Settings to test video. When the video is started it displays it to your screen and broadcasts the video to the WalkTheWeb Hub server and your local instance. 

New Streaming Video 3D Web Object - When editing a 3D Community, 3D Building, or 3D Thing, we added a new streaming video plane. Currently this displays your test video stream in the 3D Scene. When the streaming video is fully active, it will display the video stream from the WalkTheWeb Hub that you broadcast from your Admin instance.

Addition of Streaming JavaScript support - The following streaming video support scripts have been added to the core - scripts - engine libraries: recordrtc.js, recordrtc-min.js, and socket.io-stream.js. These scripts help enable your video player and microphone (you will be prompted for permission first) so that we can pick up the feed and broadcast your video stream. 

Added a wtw_streaming div to Main Elements - The wtw_streaming div is designed to be the container for on-the-fly additions of video and audio feeds using HTML 5 elements. For example, to create or receive a video feed, Babylon uses a hidden HTML 5 element for video then uses that HTML 5 element as the source for the 3D Scene video. When it is complete, the HTML 5 element is removed from the page. There will be a separate video or audio HTML 5 element for each video or sound added to the 3D Scene.

New channel on Hub for Voice Chat - We added a new channel on the WalkTheWeb Hub to support voice chat and video streaming. It is in the preliminary stages and will be released as it is completed and tested. This also added a local JavaScript file called voicechat.js as the interface to the new WalkTheWeb Hub channel. By using a separate channel we can isolate the usage from the other admin, multiplayer, and chat functionality.