Updated 3.4.2
wtwplugins PHP Class Corrections - In the returnCoveringDefsList function, the variable for $zactionzones was corrected to $zcoveringdef for the jsfunction property. Throughout the file a number of serror path references were set to class_wtwmenus which was the former location of the functions. These have been updated to reference class_wtwplugins.