Updated 3.4.2
Added Fields to actionzones Table - In order to support various uses of Action Zones, we added 2 decimal values named value1 and value2. They are also added to the actionzone object definition, connections, handlers, forms (as hidden fields by default), and functions. They can be used to store various numerical settings that can be retrieved as part of the Action Zone Definition. 

Added defaulteditform Field to actionzones Table - I needed a way to determine if the Action Zone uses the Default Edit Form for Action Zones or if it has a custom form for editing. By adding the field defaulteditform with a default value of 0, now we can set up which Action Zones use the default form and therefore show up in the edit Action Zone Drop Down list, or should not show there and will be handled in a different way.

Implemented defaulteditform in Action Zone Functions - Using the new defaulteditform field, we set the dropdown list for new Action Zones and edit existing Action Zones to only include Action Zones with the defaulteditform value that is zero. Custom Action Zones can use any other integer to set themselves apart. In addition, the class wtwplugins where it collects action zone definitions for plugins to be added to the lists, now includes a passed variable for defaulteditform.  

Open Action Zone parentname - When opening an Action Zone for editing, the parent name is set to the connecting grid name, then set to the WTW.getParentName function which allows for the connecting grid and/or another Action Zone as the parent. Since both were there and the function is the final result, it was not necessary to set it the first time, therefore removed.