Updated 3.4.3
Added Content Rating to Websites - Extended the new Admin Menu item for Requirements and renamed it Ratings and Requirements. We also added it to the 3D Communities, 3D Buildings, and 3D Things under the Options and Settings menu. You can now set each 3D Community, 3D Building, and 3D Thing with a content rating. The rating system includes: Web-All (All Visitors - Safe for All Ages.), Web-P (Parental Oversight - Adult supervision suggested for Children.), Web-P13 (Parental Caution for Children - Not recommended for Children under 13 Years Old.), Web-P17 (Parental Oversight for Visitors - Adult supervision recommended for Visitors under 18 Years Old, not recommended for Children under 13 Years Old.), and Web-Adult (Adult - Visitors must be at least 18 Years Old.). You can also add an additional Content Warning Text to the notice. 

Content Rating Listing - On the bottom right of the website Browse Menu, we added a Content Rating link that when clicked, you will see a complete content breakdown for the current 3D Website. The rating shown in the menu is the general overall rating based on all of the parts. When you click it, the breakdown will show each of the parts and their respective rating. It will also identify with an Asterisk if there are parts that are not yet rated.