Updated 3.4.3
Set First Building Cancel Process - In the 3D Community Options and Settings, under Set First 3D Building, if you selected to cancel, the 3D Building Marker in the 3D Scene was not being removed. We corrected the close function to include the removal of the 3D Building Marker (named: firstbuilding-----babylonfile).

Added 3D Community Requirements Page - In Admin under the Options and Settings for a 3D Community, we added a new page form for any requirement settings that will be specific to a 3D Community. For example, this page will set the allowed Avatar Groups, perennial warnings if necessary (Work in Progress), and 3D Plugins required for a 3D Community. 

Slider Switch CSS moved to Main CSS - The Slider, Switch, round, enablelabel, and disabledlabel CSS definitions have been moved to the /core/styles/wtw_core.css file. References to wtw3dinternet- versions of that class have been updated to wtw-. This allows it to be a standard for use in browse menus, Head up displays, and admin menus.

Corrected New User Process - When creating a new WalkTheWeb global user, there was an error that was preventing the process from completing. We corrected the issue which was preventing the wtwhandlers class from loading. It is working now.

New Content Ratings Table Added - We added a new table to cover storing the content ratings that can be used to show warnings on the 3D Websites. It is designed to store 3D Community, 3D Building, and 3D Thing designer suggested website ratings. 

Added Content Rating to Websites - Extended the new Admin Menu item for Requirements and renamed it Ratings and Requirements. We also added it to the 3D Communities, 3D Buildings, and 3D Things under the Options and Settings menu. You can now set each 3D Community, 3D Building, and 3D Thing with a content rating. The rating system includes: Web-All (All Visitors - Safe for All Ages.), Web-P (Parental Oversight - Adult supervision suggested for Children.), Web-P13 (Parental Caution for Children - Not recommended for Children under 13 Years Old.), Web-P17 (Parental Oversight for Visitors - Adult supervision recommended for Visitors under 18 Years Old, not recommended for Children under 13 Years Old.), and Web-Adult (Adult - Visitors must be at least 18 Years Old.). You can also add an additional Content Warning Text to the notice. 

Content Rating Listing - On the bottom right of the website Browse Menu, we added a Content Rating link that when clicked, you will see a complete content breakdown for the current 3D Website. The rating shown in the menu is the general overall rating based on all of the parts. When you click it, the breakdown will show each of the parts and their respective rating. It will also identify with an Asterisk if there are parts that are not yet rated.

New Ratings Value to Text Conversion Function - We added a PHP function for getRatingText into the classes for wtwdb, connect, and handlers. If you pass the rating, it will provide the rating text description.