Updated 3.4.3
Enhanced Connect Avatar Definition - Under the /connect/avatar.php file, we added the avatar parts (avatar part colors) and avatar animation definitions. This is the definitions of new avatars when you select your avatar and does not effect any currently assigned user avatars.

Ability to Edit the Avatar Definitions - We added the ability to edit the Avatar Definitions that are opened from Edit button links from the Avatar List or in the Admin Menu for 3D Avatars, Avatar Profiles. This page allows you to edit the basic information about the avatars, including the ability to delete an avatar from the available avatars to select when creating an account. There are sections for avatar parts (showing colors) and avatar animations. In the future we will make it where you can edit these colors and animations.

Avatar List Expanded - The Avatar List in Admin under 3D Avatars now shows additional information for Folder, File, Edit Link, and the ability to add a new Avatar Definition per Group.

Content Ratings now supported in Shared 3D Webs - When you share or download a 3D Web (3D Community, 3D Building, or 3D Thing), it now includes the content rating that was recommended by the 3D Web creator. This allows us to show the recommended content rating when we download an item.