Updated 3.4.3
Before the content rating menu item is added by admin, we needed to have the function check to see if the wtw_ratings id was found. If not, it skips the process to add the rating. This covers the time before the admin performs the update, or if the menu item gets manually removed.