Updated 3.4.4
New User Avatar was not Replacing Deleted User Avatar - When selecting a new User Avatar, the process checks to see if you already have an avatar with the same avatarid. If you do, it loads that avatar instead of creating a new one. The process did not account for the condition that finds a User Avatar in the table, but it is flagged as deleted. This condition should allow the creation of the new avatar. We updated the sql statement to include deleted=0 when checking.