Updated 3.4.4
Sorted the Avatar Edit Animations List - When editing a 3D Avatar, the animations list uses a load priority to keep the list in a logical order of use. This helps us load animations like Walk before less priority optional animations. We chose to order the animations within the load priority (like all of the optional ones) by the animation friendly name. This way the designer can name them in a way that effects the order. The change was made in the connect/avatar.php file but effects many pulled lists of animations in JavaScript.