Updated 3.4.4
User Avatar Animation Play section - The User animation Play icon on the center bottom on the browse screen opens a menu of play icons for each Optional animation loaded. We restyled the box, made it wider, added the names of the animations for fast reference, and removed the edit animations icon on the end (no longer used). We also removed the word Option for any Animation Friendly Name to shorten the viewed name. 

Animation Fight Mode! - The User animation Play icon on the center bottom on the browse screen opens a menu of play icons for each Optional animation loaded. If you prefaced the name with Fight, they are no longer shown by default. We added a Mode button to select between Normal and Fight. When you select Fight, the fight animations all show in the box in alphabetical order before the Options and the Avatar changes to the Fight Idle animation (onwait-fight). If you do not have a onwait-fight animation loaded, the button to change to fight mode will not appear.