Updated 3.4.4
Limit List of Avatars - The list of avatars from connect/avatars.php was including deleted when it retrieves your own (group=my). I added deleted=0 to the where clause to pull only active avatars.

New User Avatar was not Replacing Deleted User Avatar - When selecting a new User Avatar, the process checks to see if you already have an avatar with the same avatarid. If you do, it loads that avatar instead of creating a new one. The process did not account for the condition that finds a User Avatar in the table, but it is flagged as deleted. This condition should allow the creation of the new avatar. We updated the sql statement to include deleted=0 when checking.

Avatar Animations Start Frame - When loading animations to avatars, the total frames was being set to the end frame. This did not account for an animation that was not zero based. We adjusted the calculation to take the end frame minus the start frame. Now the code will support importing animations that are not zero based.

Added Increment Weight function to Test Animations - When using an avatar the animations are weighted and have an incremental transition from wait state to full animation. When we build new avatars and add animations, the test function now has a similar incremental weight applied so that it closer resembles the production use. We should be able to see relatively the same response time to transition animations. 

Added WTW.editAvatar Global Variable - When creating an Avatar, there are times when you need to test or check parts of the Avatar, skeleton, and animations using the console log. For this purpose, we added a global variable WTW.editAvatar which is the zavatar of the edit avatar. This is equivalent to the WTW.myAvatar global variable for your current avatar.

Sorted the Avatar Edit Animations List - When editing a 3D Avatar, the animations list uses a load priority to keep the list in a logical order of use. This helps us load animations like Walk before less priority optional animations. We chose to order the animations within the load priority (like all of the optional ones) by the animation friendly name. This way the designer can name them in a way that effects the order. The change was made in the connect/avatar.php file but effects many pulled lists of animations in JavaScript. 

User Avatar Animation Play section - The User animation Play icon on the center bottom on the browse screen opens a menu of play icons for each Optional animation loaded. We restyled the box, made it wider, added the names of the animations for fast reference, and removed the edit animations icon on the end (no longer used). We also removed the word Option for any Animation Friendly Name to shorten the viewed name. 

Animation Fight Mode! - The User animation Play icon on the center bottom on the browse screen opens a menu of play icons for each Optional animation loaded. If you prefaced the name with Fight, they are no longer shown by default. We added a Mode button to select between Normal and Fight. When you select Fight, the fight animations all show in the box in alphabetical order before the Options and the Avatar changes to the Fight Idle animation (onwait-fight). If you do not have a onwait-fight animation loaded, the button to change to fight mode will not appear.