Updated 3.4.4
Shortened notation for Action Zone Type - Created a variable for zactionzonetype in the setMoldActionZoneParent and transformPosition functions so that it does not have to keep referring back to the global array of Action Zones for the type value. This should slightly help performance.

Fixed the Rotate Action Zone - The rotate Action Zone was not working. The pole was not showing during edit, there was no rotation, and if you selected an object to add to the rotation, it would disappear. It turns out that when the rotation was added the variable calculation for rotation speed was named improperly, causing it to not load making the pole undefined and not visible. When the speed variable was corrected, the rotate Action Zone is now fully functional.