Updated 3.4.4
Walk and Turn Speeds added to User Avatar Table - The walk speed, walk animation speed, turn speed, and turn animation speed are saved as cookies. But for expansion we want it to transfer to the global avatars which are only stored in tables. Therefore we added the 4 columns to the useravatars table.