Updated 3.4.4
Fixed Snapshot Page - When we added the 3D Avatar Snapshot functionality, it broke the 3D Community, 3D Building, and 3D Thing process. It was updating the snapshot, but it was not showing it on the snapshot form. Only the share form was updating. Now both forms are updating as expected. Some of the code was under a conditional statement for Avatars, but needed to be in line for both conditions. It is corrected and fully functional.