Updated 3.4.4
Fixed 3D Store Buttons - Corrected a number of issues with the store buttons including start direction when you add the button to the 3D Scene and replaced the old box View Cart button with the newer oval button with the border that you can set the color.

Corrected Product Display Text Size - In the 3D Store Plugin, the product displays have text on the titles and prices. They were not scaling correctly when the product display was set to uneven values. Since planes only have 2 directions of scaling, the values applied were incorrect. We corrected the scaling in the plane settings for the effected text y swapping the lenz for leny.