Updated 3.4.4
Downloads were missing Description - When you downloaded a 3D Community, 3D Building, or 3D Thing from WalkTheWeb Downloads in the Media Library, the description was dropped from the new item. This was since the description was separated from Shared Description a couple months ago. We corrected the issue. It will now include the description as expected.