Updated 3.4.4
New Heads Up Display (HUD) - We are creating a new Heads Up Display (HUD). When completed, this will replace the bottom black menu bar. It is now created in 3D in the actual 3D Scene so it will render properly, even in VR! This view is the Main Menu. We will be adding more Menu Items as we go!

New HUD Settings Menu - We have dynamically created the new HUD Settings menu in 3D. When completed, there will be a lot more menu items, a scroll bar, and some cool animation effects.

New HUD Settings Profile Page - This is more than just a cool profile page, when Profile is selected in the menu, the menu frame stretches open, sliding the menu to the left, then swinging it forward, as the page content opens. This is a fully working 3D form  with button hovers that rotate the button, text box hovers that resize the text box slightly, Save and Cancel functions, select a text box to get a blinking 3D cursor, and 3D Text that fills as you type. The frame and menu resize using animation to go back to the settings menu display when Saved or Cancel is clicked.

New HUD Inventory Display - Even though Inventory of collected items is not in place yet, we are planning a cool menu that is morphed using animation from the Main Menu. This will be where it displays the 3D items you collect and are part of your inventory that you can use. More on that part as it progresses. For now, it is all about the HUD menu displays with its shapes and sizes; all in 3D and VR ready!

HUD Call Keys - The HUD is now opened and closed using the H or M keys on the keyboard. Additional methods for gamepad will be added after it is complete. The HUD is a beginning work in progress with a couple of menu items turned on. More will be added as they are completed.

Added the Initial HUD Menu Items - Since we created the basis for a HUD and are working to set it up, we are adding the initial HUD menu items to be shown. They have been scripted in the PHP wtwtables class. This increments the Database version to allow the data check process.