Updated 3.4.4
Added Flag for Mold Loaded - With the changes to loading 3D Communities first, we needed to change the request flag from when the mesh was called to be built, to when it was completed. To do this, we added a flag that sets the loaded = 0 and changed it to 1 when the mesh is loaded. This helps account for queue time and asynchronous loading. This effects the object definitions, mold connect files, and set shown molds functions. 

Corrected WalkTheWeb Downloads Issue - The WalkTheWeb downloads page had a conditional statement to select avatar, community, building, or thing. The statement was missing a double equals == so the statement was not operating as intended. We corrected the if conditional statement and it is working once again.