Updated 3.4.5
Fixed the Video Player - We completely reworked the video player! First we cleaned up the design in Blender and added node based coloring, fixed the animation on the buttons, orientation for new video players to be facing you, changed the scaling to be simple values (before it wanted you to keep it to the right scale ratio for the screen shape), hooked up the screen play and pause, fixed the video poster, and added the Transformation Node based Scaling Guides.