Updated 3.4.5
Web Text Field in Molds - The webtext field in the molds form was only saving under the condition if the shape was 3dtext. Since we are now using this field for 3D Store Signs and other uses, we set the condition to always save any value put into the webtext field. It has already been set to update in the setNewMolds function which changes the appearance during editing.

Deactivated Multiplayer in Admin Mode - In order to give more resources to building 3D Websites, we disabled the multiplayer in Admin Mode.

Table Fields for Versions - Since 3D Internet is constantly evolving, we wanted to add the ability to version 3D Communities, 3D Buildings, 3D Things, and 3D Avatars (all 3D Webs). This will allow us to track and provide options to update each of these 3D Webs. So as 3D Avatars get more animations, 3D Buildings change graphics, and 3D Community Scenes improve load speeds, we plan to make it easy to see if there are updates available and allow you to click to update these 3D Webs. At this stage, we added 3 fields for versionXXXXid, version, and versiondesc to each of the tables and populated the id with the key id and set the initial version to 1.0.0 for all 3D Webs. 

3D Store Molds Selectable in Admin Mode - Some of the 3D Store Molds create child meshes when loaded. In admin mode we needed to make sure you could select the Mold for edit. Therefore, we added the isPickable flag and registerMouseOver to the various relevant child parts to make it easier to select. 

Unindexed Mesh Issue - Overall the unindexed mesh setting on meshes seems to help speed up the render cycle. But there are a couple simple meshes that did not render correctly after the change. Therefore we removed the property to unindex the mesh. For example, the simple box was missing a corner. after the change and is now reversed. 

Fixed the Video Player - We completely reworked the video player! First we cleaned up the design in Blender and added node based coloring, fixed the animation on the buttons, orientation for new video players to be facing you, changed the scaling to be simple values (before it wanted you to keep it to the right scale ratio for the screen shape), hooked up the screen play and pause, fixed the video poster, and added the Transformation Node based Scaling Guides. 

Default Colors on Textures - In case the texture settings are not passed to the covering (material) step in adding a mold (mesh), we added default colors for diffuse, specular, ambient, and emissive colors and attempt to read any added values. 

3D Models new Admin Mode Settings - When adding a 3D Model to a 3D Scene, we were limited to the settings that came from the 3D Model when it was built. We added new settings to assist in overriding or complimenting the 3D Model. These settings are diffuse color, specular color, emissive color, ambient color, receive shadows, and add water reflections. The receive shadows setting can assist if the 3D Model has shadows that make them too dark to see the details. Note that some of the color settings will not modify some of the 3D Models depending on the applied nodes, textures, and materials. These settings may also effect some of the already deployed 3D Models since we do set default colors if none are found. Please recheck your 3D Models coloring in your 3D Scenes as needed. 

Admin Mode Edit Mold Water Reflection - The water reflection setting was not being set when you open a mold for edit. We corrected the load form code and also implemented the setting as part of the setNewMold function that is used to live update the mold as it is being edited.

Enabled Reflections - In scenes with water, we once again enabled the use of reflections and refractions. The related functions now use the latest Water Material. We modified the add function and created a remove function. We also added the function calls to the babylonfile 3D Models mold type, Avatars, and disposeClean functions.