Updated 3.4.5
New 3D Compass - We are working replace the Browse Menu with the new HUD. We just replaced the old compass with a new 3D Compass that is in the 3D Scene. It attaches to the active camera so it will always be in front of the user camera. It will rotate to the correct angle of the avatar movement, zero degrees being North. In addition, it has a Red Arrow that will always point to the closest 3D Building in the 3D Scene (really helps when you wander off into the distance).

Toggle Compass - We replaced the toggleCompass function with hudToggleCompass in the wtw_hud.js file. This function loads the compass and parents it to the new camerafront transform node that stays 20 units in front of the current active camera. The new Babylon files, textures, and Blender files for the compass can be found in the /content/system/babylon/compass folder. You can press c on the keyboard to toggle the compass on-off.